+++ Cluster/src/main/resources/sparkJob.properties
@@ -31,7 +31,7 @@ job.faceObjectConsumer.appName=FaceObjectConsumer
 job.faceObjectConsumer.broker.list=
 job.faceObjectConsumer.group.id=FaceObjectConsumerGroup
 job.faceObjectConsumer.topic.name=feature
-job.zkDirAndPort=
+job.zkDirAndPort=172.18.18.103:2181,172.18.18.104:2181,172.18.18.105:2181
 job.kafkaToEs.zkPaths=/es
 job.kafkaToParquet.zkPaths=/parquet
 job.repartition.number=3
